+++ test/call_new_api.rb
@@ -13,11 +13,11 @@ class CallTest < Minitest::Spec
 
   let(:form) { SongForm.new(Song.new) }
 
-  it { form.(title: "True North").success?.must_equal true }
-  it { form.(title: "True North").failure?.must_equal false }
-  it { form.(title: "").success?.must_equal false }
-  it { form.(title: "").failure?.must_equal true }
+  it { _(form.(title: "True North").success?).must_equal true }
+  it { _(form.(title: "True North").failure?).must_equal false }
+  it { _(form.(title: "").success?).must_equal false }
+  it { _(form.(title: "").failure?).must_equal true }
 
-  it { form.(title: "True North").errors.messages.must_equal({}) }
-  it { form.(title: "").errors.messages.must_equal(title: ["must be filled"]) }
+  it { _(form.(title: "True North").errors.messages).must_equal({}) }
+  it { _(form.(title: "").errors.messages).must_equal(title: ["must be filled"]) }
 end