+++ test/coercion_test.rb
@@ -47,9 +47,9 @@ class CoercionTest < BaseTest
   end
 
   # it { subject.released_at.must_be_kind_of DateTime }
-  it { subject.released_at.must_equal "31/03/1981" } # NO coercion in setup.
-  it { subject.hit.length.must_equal "312" }
-  it { subject.band.label.value.must_equal "9999.99" }
+  it { _(subject.released_at).must_equal "31/03/1981" } # NO coercion in setup.
+  it { _(subject.hit.length).must_equal "312" }
+  it { _(subject.band.label.value).must_equal "9999.99" }
 
   let(:params) do
     {
@@ -75,24 +75,24 @@ class CoercionTest < BaseTest
   describe "#validate" do
     before { subject.validate(params) }
 
-    it { subject.released_at.must_equal DateTime.parse("30/03/1981") }
-    it { subject.hit.length.must_equal 312 }
-    it { subject.hit.good.must_equal false }
-    it { subject.band.label.value.must_equal "9999.999999.99" } # coercion happened once.
-    it { subject.metadata.publication_settings.featured.must_equal false }
+    it { _(subject.released_at).must_equal DateTime.parse("30/03/1981") }
+    it { _(subject.hit.length).must_equal 312 }
+    it { _(subject.hit.good).must_equal false }
+    it { _(subject.band.label.value).must_equal "9999.999999.99" } # coercion happened once.
+    it { _(subject.metadata.publication_settings.featured).must_equal false }
   end
 
   # sync
   describe "#sync" do
     before do
-      subject.validate(params).must_equal true
+      _(subject.validate(params)).must_equal true
       subject.sync
     end
 
-    it { album.released_at.must_equal DateTime.parse("30/03/1981") }
-    it { album.hit.length.must_equal 312 }
-    it { album.hit.good.must_equal false }
+    it { _(album.released_at).must_equal DateTime.parse("30/03/1981") }
+    it { _(album.hit.length).must_equal 312 }
+    it { _(album.hit.good).must_equal false }
     it { assert_nil album.metadata[:publication_settings] }
-    it { album.metadata["publication_settings"]["featured"].must_equal false }
+    it { _(album.metadata["publication_settings"]["featured"]).must_equal false }
   end
 end